+++ packages/components/src/components/option-group/option-group.browser.e2e.tsx
@@ -1,6 +1,8 @@
-import { describe } from "vitest";
+import { describe, expect, it } from "vitest";
+import { page } from "vitest/browser";
+import { h } from "@arcgis/lumina";
 import { mount } from "@arcgis/lumina-compiler/testing";
-import { defaults, reflects, hidden, renders } from "../../tests/commonTests/browser";
+import { defaults, hidden, reflects, renders } from "../../tests/commonTests/browser";
 
 describe("calcite-option-group", () => {
   describe("defaults", () => {
@@ -34,4 +36,11 @@ describe("calcite-option-group", () => {
   describe("renders", () => {
     renders(() => mount("calcite-option-group"), { display: "inline", visible: false });
   });
+});
+
+it("has a label", async () => {
+  await mount(<calcite-option-group label="test-group" />);
+  const label = page.getByText("test-group");
+
+  await expect.element(label).toBeVisible();
 });